Functional 2 (#5)
* Misc remove redundant bindings

* Refactor variable code

* Improvements to the obj loader

* Add dragon mesh

* More obj loader related changes

* Misc improvements to building and testing

* Add more meshes to test with

* Add demo to just load a mesh

* Fix bug in obj loader

* Add support for marshalling static lists from python

* Improve support for marshalling points from python to cpp

* Misc improvement attempts

* Misc changes

* First successful test of serialization

* Move the type info into the core

* Remove redundant file and misc changes

+++ include/galcore/ObjLoader.h
@@ -12,9 +12,9 @@ class ObjMeshData
 public:
   struct Face
   {
-    uint32_t vertices[3];
-    uint32_t texCoords[3];
-    uint32_t normals[3];
+    int64_t vertices[3];
+    int64_t texCoords[3];
+    int64_t normals[3];
   };
 
   ObjMeshData(const std::filesystem::path& filepath, bool flipYAndZ = false);
